cleanup(pxarmount): remove dead code, normalize naming, consolidate utilities
