--- internal/pxarmount/xattr.go
@@ -1,3 +1,0 @@
-package pxarmount
-
-// xattr helpers — currently empty, reserved for future ACL support.
